Fix components composition
