+++ client/index.html
@@ -53,7 +53,7 @@
         <p class="Form_loader">Loading…</p>
       </div>
 
-      <div class="Form_messages" data-id="alert_messages"></div>
+      <div class="Form_messages" data-id="form_messages"></div>
 
       <div class="Form_coverage" data-id="form_coverage" hidden>
         <h2 class="Interactive_title">Audience coverage: <span data-id="form_total"></span></h2>
